test/Microsoft.IdentityModel.Xml.Tests/SignatureTests.cs
//------------------------------------------------------------------------------
//
// Copyright (c) Microsoft Corporation.
// All rights reserved.
//
// This code is licensed under the MIT License.
//
// Permission is hereby granted, free of charge, to any person obtaining a copy
// of this software and associated documentation files(the "Software"), to deal
// in the Software without restriction, including without limitation the rights
// to use, copy, modify, merge, publish, distribute, sublicense, and / or sell
// copies of the Software, and to permit persons to whom the Software is
// furnished to do so, subject to the following conditions :
//
// The above copyright notice and this permission notice shall be included in
// all copies or substantial portions of the Software.
//
// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT.IN NO EVENT SHALL THE
// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
// THE SOFTWARE.
//
//------------------------------------------------------------------------------

using System;
using System.Collections.Generic;
using System.Security.Cryptography;
using Microsoft.IdentityModel.Tests;
using Microsoft.IdentityModel.Tokens;
using Microsoft.IdentityModel.Xml;
using Xunit;

#pragma warning disable CS3016 // Arrays as attribute arguments is not CLS-compliant

namespace Microsoft.IdentityModel.Xml.Tests
{
    public class SignatureTests
    {
        [Fact]
        public void GetSets()
        {
            var type = typeof(Signature);
            var properties = type.GetProperties();
            Assert.True(properties.Length == 5, $"Number of properties has changed from 5 to: {properties.Length}, adjust tests");
            var context = new GetSetContext
            {
                PropertyNamesAndSetGetValue = new List<KeyValuePair<string, List<object>>>
                {
                    new KeyValuePair<string, List<object>>("Id", new List<object>{(string)null, Guid.NewGuid().ToString(), Guid.NewGuid().ToString()}),
                    new KeyValuePair<string, List<object>>("KeyInfo", new List<object>{(KeyInfo)null, new KeyInfo(), new KeyInfo()}),
                    new KeyValuePair<string, List<object>>("Prefix", new List<object>{"", Guid.NewGuid().ToString(), Guid.NewGuid().ToString()}),
                    new KeyValuePair<string, List<object>>("SignatureValue", new List<object>{(string)null, Guid.NewGuid().ToString(), Guid.NewGuid().ToString()}),
                    new KeyValuePair<string, List<object>>("SignedInfo", new List<object>{(SignedInfo)null, new SignedInfo(), new SignedInfo()}),
                },
                Object = new Signature(),
            };

            TestUtilities.GetSet(context);
            TestUtilities.AssertFailIfErrors($"{this}.GetSets", context.Errors);
        }

        [Fact]
        public void Publics( )
        {
            var signature = new Signature();
            var properties = new List<string>()
            {
                "SignatureValue",
                "SignedInfo",
            };

            var context = new GetSetContext( );
            foreach (string property in properties)
            {
                TestUtilities.SetGet(signature, property, null, ExpectedException.ArgumentNullException(substringExpected: "value"), context);
            }

            TestUtilities.AssertFailIfErrors($"{this}.Publics", context.Errors);
        }

        [Theory, MemberData(nameof(ConstructorTheoryData))]
        public void Constructor(SignatureTheoryData theoryData)
        {
            TestUtilities.WriteHeader($"{this}.Constructor", theoryData);
            var context = new CompareContext($"{this}.Constructor, {theoryData.TestId}");
            try
            {
                var signature = new Signature(theoryData.SignedInfo);
                theoryData.ExpectedException.ProcessNoException(context);
            }
            catch (Exception ex)
            {
                theoryData.ExpectedException.ProcessException(ex, context);
            }

            TestUtilities.AssertFailIfErrors(context);
        }

        public static TheoryData<SignatureTheoryData> ConstructorTheoryData
        {
            get
            {
                return new TheoryData<SignatureTheoryData>()
                {
                    new SignatureTheoryData
                    {
                        ExpectedException = ExpectedException.ArgumentNullException("IDX10000: The parameter 'value' cannot be a 'null' or an empty object."),
                        First = true,
                        SignedInfo = null,
                        TestId = "SignedInfo NULL"
                    },
                    new SignatureTheoryData
                    {
                        SignedInfo = new SignedInfo(),
                        TestId = "SignedInfo"
                    }
                };
            }
        }

        [Theory, MemberData(nameof(VerifyTheoryData))]
        public void Verify( SignatureTheoryData theoryData )
        {
            var context = TestUtilities.WriteHeader($"{this}.Verify", theoryData);
            try
            {
                theoryData.Signature.Verify(theoryData.SecurityKey);
            }
            catch (Exception ex)
            {
                theoryData.ExpectedException.ProcessException(ex, context);
            }

            TestUtilities.AssertFailIfErrors(context);
        }

        public static TheoryData<SignatureTheoryData> VerifyTheoryData
        {
            get
            {
                var key = Default.AsymmetricSigningKey;
                var cryptoProviderFactory = new CustomCryptoProviderFactory(new string[] { SecurityAlgorithms.RsaSha256Signature });
                key.CryptoProviderFactory = cryptoProviderFactory;

                var keyUnknownDigest = Default.AsymmetricSigningKey;
                cryptoProviderFactory = new CustomCryptoProviderFactory(new string[] { SecurityAlgorithms.RsaSha256Signature })
                {
                    SignatureProvider = new CustomSignatureProvider(key, SecurityAlgorithms.RsaSha256Signature)
                };

                keyUnknownDigest.CryptoProviderFactory = cryptoProviderFactory;

                var signatureUnknownReferenceDigestAlg = Default.Signature;
                signatureUnknownReferenceDigestAlg.SignedInfo.References[0].DigestMethod = $"_{SecurityAlgorithms.Sha256Digest}";

                return new TheoryData<SignatureTheoryData>
                {
                    new SignatureTheoryData
                    {
                        ExpectedException = new ExpectedException(typeof(XmlValidationException), "IDX30212:"),
                        First = true,
                        SecurityKey = Default.AsymmetricSigningKey,
                        Signature = new Signature(),
                        TestId = "SignedInfo:Null"
                    },
                    new SignatureTheoryData
                    {
                        ExpectedException = ExpectedException.ArgumentNullException("key"),
                        SecurityKey = null,
                        Signature = new Signature(),
                        TestId = "SecurityKey:Null"
                    },
                    new SignatureTheoryData
                    {
                        ExpectedException = new ExpectedException(typeof(XmlValidationException), "IDX30207:"),
                        SecurityKey = Default.AsymmetricSigningKey,
                        Signature = new Signature(new SignedInfo{SignatureMethod = SecurityAlgorithms.Aes128CbcHmacSha256 }),
                        TestId = "Signature:MethodNotSupported"
                    },
                    new SignatureTheoryData
                    {
                        ExpectedException = new ExpectedException(typeof(XmlValidationException), "IDX30203:"),
                        SecurityKey = key,
                        Signature = new Signature(new SignedInfo{SignatureMethod = SecurityAlgorithms.RsaSha256Signature }),
                        TestId = "SignatureProvider.CreateForVerifying:ReturnsNull"
                    },
                    new SignatureTheoryData
                    {
                        ExpectedException = new ExpectedException(typeof(XmlValidationException), "IDX30208:"),
                        SecurityKey = keyUnknownDigest,
                        Signature = signatureUnknownReferenceDigestAlg,
                        TestId = "Reference:UnknownDigestAlg",
                    }
                };
            }
        }

        private static SignatureTheoryData SignatureTest(SignatureTestSet testSet, SecurityKey key, XmlTokenStream tokenStream, ExpectedException expectedException = null, bool first = false)
        {
            return new SignatureTheoryData
            {
                ExpectedException = expectedException ?? ExpectedException.NoExceptionExpected,
                SecurityKey = key,
                Signature = testSet.Signature,
                TestId = testSet.TestId ?? nameof(testSet),
                TokenStream = tokenStream,
                Xml = testSet.Xml
            };
        }
    }

    public class SignatureTheoryData : TheoryDataBase
    {
        public SecurityKey SecurityKey
        {
            get;
            set;
        } = Default.AsymmetricSigningKey;

        public DSigSerializer Serializer
        {
            get;
            set;
        } = new DSigSerializer();

        public Signature Signature
        {
            get;
            set;
        }

        public SignedInfo SignedInfo
        {
            get;
            set;
        }

        public XmlTokenStream TokenStream
        {
            get;
            set;
        } = Default.TokenStream;

        public string Xml
        {
            get;
            set;
        }
    }
}

#pragma warning restore CS3016 // Arrays as attribute arguments is not CLS-compliant
